rename keep-alive.ts -> KeepAlive.ts
